hoàn thành chức năng phiếu gửi tiền

+++ src/main/java/com/example/sweet/controller/GiaoDich/KenhGiaoDichController.java
@@ -26,7 +26,7 @@ public class KenhGiaoDichController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteKenhGiaoDich(@PathVariable int id) {
+    public void deleteKenhGiaoDich(@PathVariable Long id) {
         respository.deleteById(id);
     }
 }
